+++ src/Graphql/Client.php
@@ -16,7 +16,7 @@ class Client
     public function request(string $query, array $variables = [])
     {
         $response = Http::withToken($this->token)
-            ->withCorrelationId()
+            ->withCorrelation()
             ->timeout($this->timeout)
             ->post($this->url, compact('query', 'variables'))
             ->throw();